Split into villager & dolphin manager
